Entity version (#62)
* added Version field on Entity

* added Version to Entity.Equals check

* added Version field on Entity

* added Version to Entity.Equals check

* small change on Entity.IsAlive

* should fix ci while multi dotnet setup is possible

+++ source/DefaultEcs/EntitySet.cs
@@ -20,7 +20,7 @@ namespace DefaultEcs
         #region Fields
 
         private readonly bool _needClearing;
-        private readonly int _worldId;
+        private readonly short _worldId;
         private readonly int _maxEntityCount;
         private readonly Predicate<ComponentEnum> _filter;
         private readonly IDisposable _subscriptions;
